testingdgcnn/requirements.txt
torch==2.0.1
torchvision==0.15.2
numpy==1.26.4
opencv-python==4.8.1.78
scipy==1.11.4
pillow==10.1.0
scikit-learn==1.3.2
matplotlib==3.8.2
open3d==0.17.0
timm==0.9.12
tqdm==4.66.1
tensorboard==2.15.1
requests==2.31.0
